fix syncgroup builtin for K1 MPPA
syncgroup takes a 64bits operand instead of the previous 32bits used by older MPPA version.
ref T7376

	* gcc/config/k1/builtin_k1.h (__builtin_k1_syncgroup): Change
	syncgroup insn operand type to 64bits.
	* gcc/config/k1/k1.c (k1_target_init_builtins): Likewise.
	(k1_expand_builtin_syncgroup): Likewise.
	* gcc/config/k1/k1.md (syncgroup): Likewise.

+++ gcc/config/k1/builtin_k1.h
@@ -82,7 +82,7 @@ extern void __builtin_k1_sleep (void);
 /* STOP */
 extern void __builtin_k1_stop (void);
 /* SYNCGROUP */
-extern void __builtin_k1_syncgroup (unsigned int);
+extern void __builtin_k1_syncgroup (unsigned long long);
 /* TRAPA */
 /* TRAPO */
 /* TLBDINVAL */
